+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 ## vNEXT
+### Highlights :tada:
++ Updated scalajs-dom to 1.0.0  [PR #362](https://github.com/shadaj/slinky/pull/362)
+
+### Breaking Changes :warning:
++ Due to the update of scalajs-dom to 1.0.0 a support for `dd` and `dt` tags has been dropped.
 
 ## [v0.6.5](https://slinky.dev)
 ### Highlights :tada: