+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## vNEXT
+
+## [v0.6.6](https://slinky.dev)
 ### Highlights :tada:
 + Updated `scalajs-dom` to 1.0.0  [PR #362](https://github.com/shadaj/slinky/pull/362)
 + Add facades for the `React.Profiler` component [PR #372](https://github.com/shadaj/slinky/pull/372)
@@ -9,11 +11,12 @@
 ### Bug Fixes :bug:
 + Fix compilation errors on Scala 2.13 when `Any` is involved in `Props`/`State` types [PR #388](https://github.com/shadaj/slinky/pull/388)
 + Allow exporting external component definitions as instances of `ReactComponentClass` [PR #377](https://github.com/shadaj/slinky/pull/377)
++ Update prop type definitions for the React Native `Image` component to support local images [PR #411](https://github.com/shadaj/slinky/pull/411)
 
 ### Breaking Changes :warning:
 + Due to the update of scalajs-dom to 1.0.0 a support for `dd` and `dt` tags has been dropped.
 
-## [v0.6.5](https://slinky.dev)
+## [v0.6.5](https://5e90c70b95e02b0007c114c0--slinky.netlify.app/)
 ### Highlights :tada:
 + Add docs for the new Electron app template [PR #339](https://github.com/shadaj/slinky/pull/339)